fix: :bug: Routing without modals
