+++ package.json
@@ -16,6 +16,7 @@
     "react-icons": "^4.3.1",
     "react-query": "^3.34.16",
     "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.2",
     "rxjs": "^7.5.4"
   },
   "devDependencies": {
